fix-approved supervisor request

+++ app/controllers/api/v1/expense_report_requests_controller.rb
@@ -1,7 +1,7 @@
 module Api::V1
     class ExpenseReportRequestsController < ApiController
       before_action :set_request, only: [:show]
-      skip_before_action :set_current_user_from_header, only: [:review_request]
+      skip_before_action :set_current_user_from_header, only: [:review_request, :response_request]
       skip_before_action :verify_authenticity_token
 
       def index
@@ -51,7 +51,7 @@ module Api::V1
             file: root_url + rails_blob_path(i.file, disposition: "attachment")     
           }
         end
-        if  result[:state] == "link_expired"
+        if  result[:state] == "link_expired" || result[:request].request_state.name != 'awaiting approval'
           render json: { message: "Link expired", success: true, request: result[:request], user: result[:user],request_date: result[:request_date]}, status: :ok
         else
           render json: { message: "request", success: true, invoices: invoices ,request: result[:request], user: result[:user], request_date: result[:request_date] }, status: :ok
@@ -62,13 +62,13 @@ module Api::V1
         response = params[:request][0]
         request = ExpenseReport::Request.find(params[:request][1].to_i) 
 
-        if response == 'true'
+        if response == 'true' && request.request_state.name == 'awaiting approval'
           request.update(request_state_id: ExpenseReport::RequestState.find_by(name: 'open').id)
           UserNotifierMailer.notification_new_request(request).deliver
           UserNotifierMailer.notification_request_approved_to_user(request).deliver
           UserNotifierMailer.notification_request_approved_to_boss(request).deliver
           render json: { message: "true" }, status: :ok
-        elsif response == 'false'
+        elsif response == 'false' && request.request_state.name == 'awaiting approval'
           UserNotifierMailer.notification_request_rejected(request).deliver
           UserNotifierMailer.notification_request_rejected_to_boss(request).deliver
           request.destroy
